Add spaceRenamedListener and make link redirect to space settings
