+++ webapp/portlet/src/main/java/org/exoplatform/social/portlet/SpaceMenuPortlet.java
@@ -44,7 +44,6 @@ public class SpaceMenuPortlet extends GenericPortlet {
     Locale locale = request.getLocale();
     String currentUser = request.getRemoteUser();
     try {
-      SpaceUtils.renameNavigationAndGroup(space);
       List<UserNode> navigations = SpaceUtils.getSpaceNavigations(space,
                                                                   locale,
                                                                   currentUser);
