+++ webapp/portlet/src/main/webapp/space-settings/components/SpaceSettingFormDrawer.vue
@@ -243,7 +243,8 @@ export default {
           window.setTimeout(() => {
             this.$refs.spaceFormDrawer.close();
           }, 200);
-          window.location.href = `${eXo.env.portal.context}/g/${space.groupId.replace(/\//g, ':')}`;
+
+          window.location.href = `${eXo.env.portal.context}/g/${space.groupId.replace(/\//g, ':')}/${space.displayName}/settings`;
         })
         .catch(e => {
           // eslint-disable-next-line no-console
